requirements.txt
streamlit
folium
streamlit-folium
fpdf2
requests
pandas
